Change option rendering to use jsp:include.
renderOptions() returns a string that is placed verbatim on the JSP
page. This was replaced by getOptionsPage() which returns a filename
that needs to be included using jsp:include. The group creator provider
needed to be a full JSP page since the form is a lot more complicated
than the simple string return provided by renderOptions().

Most of the existing providers and consumers just had their
renderOptions() replaced with plain html files. BlackboardProvider's was
replaced by a JSP file.

The BlackboardProvider's getGroupSet function was moved into the
BlackboardUtil helper class since it needed to be called by the
BlackboardProvider's associated JSP page.