Pressing enter in the search box initiates search.
For the GroupCreator provider:

Renamed the 'Apply' button to 'Search'.

Add JavaScript so that if people pressed enter in the search text box,
it's the same as clicking the 'Search' button.

Previously, if people pressed enter after entering a term in the search
textbox, the form gets submitted. Now, if a user instinctively presses
enter after entering their search term, it'll properly do a search
instead.