Fix GroupCreator search failing in other courses.
These changes all only apply to the GroupCreator provider.

Search wasn't working in other courses since I hardcoded the test
course's course id into it and forgot about it. Search will now properly
use the current course's course id.

It turns out that the bbNG:inventoryList's 'showAll' button doesn't
appear to work correctly after the ajax request, so I've disabled it by
making the list to show all by default.

A primitive logging mechanism was added when tracking down this bug. I
decided to keep it in. Primitive because all it does is concat log
messages to one string. A JSP page needs to get the string and print it
out for us to see it.