GroupCreator now allows multiple search criteria.
Previously limited to search on only 1 field. Multiple search is
basically implemented by stacking many single searches together.

The jstl-1.2.jar file provides an implementation of the Java Standard
Tag Library by the Apache Maven project. It was needed to provide
variable inventoryList columns. The variable columns are due to the need
to display values for all columns being searched on.

The SearchCriteria class is only meant as a convenience class to hold
all the 3 values necessary for search: The id of the field being
searched on, the operator to apply to the search, and the actual term
being searched for.

SearchFieldInfo is also a convenience class meant to hold 2 values: The
name of a field being searched and the value that field holds. It's a
Java Bean that is used to provide values for the search result table.
Each UserWrapper has a list of them holding values applicable to that
user only.