Add shortcuts to the index page.
Instead of having to select the provider and consumer manually every
time, a user can just click a shortcut for the same result. This should
be a temporary solution until we think of something better.

The shortcuts just activate a javascript function that sets the correct
options for the provider/consumer select elements and then submit the
form.

Inline CSS was used to style the shortcuts.