Add GroupCreator user info search functionality.
Enables actually searching by the user info. The old
searchSingleCriteria method was refactored out to
searchSingleCriteriaLineitem since it only deals with lineitems.

The new searchSingleCriteria first checks to see which type of search
we're doing. If we're doing a user info search, it calls the new
searchSingleCriteriaUserinfo method. If not, it calls the
searchSingleCriteriaLineitem method.

The checking for which type of search basically sees if we can match the
field id with the list of field ids known to be user info options.

The actual search is pretty much the same as searchSingleCriteria,
except we have score replaced by target.