Fix CSV Provider form submit and parsing.
The Form submit wasn't working since the the new options page form
didn't have the correct encoding type. Forms that have an file upload
element is required to have enctype='multipart/form-data'. To this end,
the Provider was extended with a method 'hasFileUpload', which returns
true if the Provider has an upload element and false otherwise.

hasFileUpload is used to determine the enctype of the options.jsp form,
if it returns true, the multipart encoding is used, if false, the
default application/x-www-form-encoded encoding is used.

There is a filePicker bbNG tag that can be used later instead for file
upload instead of the current regular html form element solution.
Unfortunately, there is almost no documentation on this tag, as usual.
Some experimentation found it quite fussy to use. Regardless, I've
converted the CSV Provider options page to a jsp in preparation for
using the filePicker tag instead since it provides built in integration
with the content collection.

CSV Provider parsing was changed to skip the very first line. Since the
CSV Consumer exports a CSV file with the first line as a header, it
didn't make sense to have the CSV Provider be unable to parse a file
generated by it's counterpart.