CSV operations use studentID instead of username.
Since the usernames are PUIDs, they're a bit harder for faculty/staff to
use than student ids. There's apparently no way for faculty/staff to
find out who exactly a PUID identifies.  So now the CSV provider and
consumer uses student IDs.

To do this, the ca.ubc.ctlt.group.User object was renamed to GroUser. It
was a bit confusing to have it conflicting with blackboard API's User.
Instead of specifying the full path, I've just renamed it. The new
GroUser object is now basically a blackboard API User wrapper.

When given a student number, it searches the course enrolment to find a
match.