CSV Provider now use bbNG:filePicker for uploads.
The filePicker is a customized version of the regular html file upload
element. One difference is that the filePicker allows you to pick files
from the content collection system.

Note that in order to use filePicker, I had to use blackboard's
MultipartRequest handler. The custom UploadMultipartRequestWrapper was
redone to become a thin wrapper for blackboard's MultipartRequest.

The original filter system was left in place, since MultipartRequest
proved to be incompatible with the whole Manager/ManagedObject system
we've got going. Mainly due to MultipartRequest being a custom object
that has no relation to the usual HttpServletRequest object.

In the CSV Provider, CSV parsing is now separated from file acquisition
with it's own helper method. This is due to there being two sources for
file acquisition: the content system and a user upload.