Fix GroupCreator bugs due to search result gen.
There are 3 bugs related to the search result generation.

When search with a single criteria, the search result table doesn't make
a column with that criteria. This is a bug resulting from when I took
the column generation code out of the searchSingle*() functions into a
helper function. I forgot to call the new helper function in the single
criteria search.

When searching with a user info criteria as a second parameter, e.g.:
First Name, the search crashes. This also results from the search result
column data generation code refactoring. It needs to ignore user info
criterias since they're already displayed to the user.

When searching using the 'or' combination operator and on two fields
which yield a result where one or more users might not have a value for
one of the fields, the search crashes. This was fixed by defaulting
fields with no values to the empty string.