GroupCreator allow column sorting on the results.
The inventoryList that displays a table of the users returned by
GroupCreator search now allows sorting on the user info field columns.

Unfortunately, due to how inventoryList implemented column sorting, I
cannot enable sorting on the grade center field columns. Apparently, you
have to pass in a Comparator for each listElement. There is a
GenericFieldComparator class specifically provided by Blackboard for
this purpose. You have to know *exactly* which fields you want to
compare, hence why it can't be used for the Grade Center columns, as
anyone can create custom columns.

Note that in order to enable sorting on the search results, I had to
cache the search criterias in session. The way inventoryList's ajax
sorting works is that it reloads the userslist.jsp page every time you
click sort. Since we can't provide the parameters being used for the
search results, when sort loads userslist.jsp, it's just the default one
with ALL the users. Caching the criterias in session works around this.

Unfortunately, when you change the sorting order on the results, the
user's selections are also forgotten. Hopefully this won't be too much
of an inconvenience.