GroupCreator filter out users who are unavailable.
Looks like students who are dropped from a course are marked as
unavailable instead of being deleted. Since there probably isn't a use
case of making groups with dropped students, I decided that it would be
fine to just not list users are marked as unavailable in the grade
center.