BlackboardConsumer option, add to existing group.
Adding users to an existing group is only available if the Provider
cannot provide multiple groups.

Adding users to an existing group just involves renaming the incoming
group from the provider to have the same title as the existing group.

Possible issue in the future: BlackboardConsumer identifies groups by
their titles, if two group has the same title, there is no way to
distinguish between them.

When you select a group in BlackboardConsumer, it will display a list of
the users already in the group. This is an ajax request similar to how
search results work in GroupCreator.

Files for the next feature, picking filename to download for the CSV
exporter, accidentally made it into this commit. They just set up the
intial starting point for it though.