CSV Consumer allow download with custom filename.
The CSV Consumer now has the option to switch between downloading the
export or saving it to the content collection system. This comment only
enables the download export part. Users can now set a custom filename to
be downloaded. The user provided filename undergoes some basic
sanitization: semicolons (;) are stripped out of filenames since it's
used for delineation in the header values and control (non-printable)
characters are stripped, as they can be used to insert new header
values.