CSV Consumer enable saving to Content Collection.
Now can export the group CSV file to the user's Content Collection. It's
basically the same process as download, except that we write the CSV to
a buffer which is then used by the content system to create a file.

The Content Collection save requires the User to select a file name and
a directory to save to. Note that if the user selects a file name that
already exists, that file will be overwritten. And if the user does not
pick a directory, it defaults to / and fails since the user probably
don't have permission.