Display errors nicely.
Change process.jsp to tell user about failures if Manager.getErrors() is
not empty. And success if it is empty.

GroupCreator, if no students were selected, it will error out and no
group will be created.

CSV Consumer, if saving to Content Collections fail, it'll show a reason
for the failure. The most common reason for failure is probably that the
user didn't select a Content Collection folder, so one of the error
messages ask if the user has selected a folder.