IE8 hacks, rename conflicting js functions names.
IE8 doesn't properly inherit CSS properties, so each provider/consumer
options doesn't have the correct left margin when rendered. This was
fixed by putting the whole options section into the IE8Hacks class and
manually specifying a left margin for that class. It's like the IE box
model fix.

Turns out that I had a few javascript functions with the same name that
can be paired up in the provider/consumer relationship. These functions
were renamed to have a unique prefix according to which provider or
consumer they belong to.