GroupCreator, better calculated columns filter.
The previous filter assumed that all columns which doesn't have Score
entries are calculated columns. This is true IF the course has at least
one student. This one student provides values for all non-calculated
columns, even if the value is empty. If the course doesn't have any
students, then non-calculated columns falls to the filter, with the
result that no columns except the user info fields are shown as options.

The fix was, of course, to use functions that Blackboard mark as "For
Internal Use Only" and doesn't bother to document.