GroupCreator, hide Name if add to existing group.
Move Instructions to the top. Also added JavaScript that detects if the
BlackboardConsumer is being used, if it is, hide the GroupName if the
user has selected to Add to an Existing Group.

Note that this is probably a bad design decision as it makes the
GroupCreator module less independent. But it was a feature request. In
the future, we should find a better way of allowing communication
between modules.