GroupCreator, warn if group name already exists.
Due to how group names being the unique identifier for groups, if the
user gives a group name that is already taken, it'll do a group merge
with the selected users instead of creating a new group. Sometimes, this
behaviour may be desired, so we'll just warn the user if they use a
group name that already exists.