BlackboardProvider group list expand view members
The group list now only shows group names. Clicking on the group name
shows a list of group members in the format of "Name (Student Number)".

The group members lists are given unique monotonically increasing IDs
and they are hidden and shown by removing a class that has display:
none.