GroupCreator only list students.
Only students have column values in the grade center, so it makes a bit
more sense that we don't list instructors in the search.

This is partially due to our switch from using the username (populated
by PUID) as the identifier to student numbers. Since instructors don't
have student numbers, they won't be recognized.