CSVProvider fix error if no GroupSet col in csv.
If the import file does not have a GroupSet column, the CSVProvider
exits as if it was successful, even though it wasn't. This was due to a
silently suppressed null pointer exception during an array index. The
fix was to refactor the code a bit so that we avoid using variables with
null values as an array index.

Also added test cases to make sure this doesn't break in the future.
Additional test cases now test that missing Username or Student ID
columns also doesn't break. And a new test that check if mixing Username
and Student IDs work.

Removed some unecessary imports in GroupCreatorProvider.