Fix permission denied error during csv import.
For some reason, our custom UBC roles doesn't like CSV import. It errors
out with a permission error but we can't figure out why. Trial and error
shows that this fixes it, but we can't logically figure out why since
this same permission request is also on every other page and every other
page seems to work fine without it.

As such, I'm entering this commit alone so it's easy to revert if we
ever do figure out what was wrong.