CsvImport: add test for 2 column imports.
We really only require 2 columns for import, the group name and some
sort of user identification. Existing tests only tests a minimum of 3
columns. Added a new test where only the group name and username is
available and a test where only the group name and student ID is
available.