BbConsumer, reduce database queries.
We've been having timeouts in huge merge courses where we might have a
few thousand students. Seeing as how the Webwork integration experienced
a similar slowdown due to database queries, I've tried to reduce the
number of database queries performed during import. The only database
read queries performed will be at the start of the import process.
Database write queries were not batched as the Blackboard API doesn't
support batched GroupMembership creation according to the documentation.

Also fixed a bug where error messages can be hidden during import if
there are multiple groups being imported. As long as it doesn't error
out on the first group, no errors will be shown. This is due to a faulty
assumption in the goodbye() method where we assume as long as there is a
previous group successfully created, there were no errors.