Convert to servlet type MVC architecture.
Instead of the index, option, and process JSP files pulling double duty
as the controller and the view, the controller code is now split off
into servlets.

This was done to fix weird bugs that cropped up after the BBL 9.1 SP11
update. For some reason, uploading a file for import breaks on redirect.
If we fix that, it breaks file export to the content collection system.
This might have something to do with changes to the default servlet
created for the the servlet-less JSP files we were using previously.
Specifying a servlet seems to have fixed these two issues.

Most of the changes are straightforward copy and pasting of Java code in
the JSP files to the servlets. File upload handling was changed a bit,
the filter has been removed and code was added to handle multipart
uploads directly in the servlet.

BlackboardProvider's user search is currently broken with this change,
needs to be fixed later.