Fix ajax operations not working.
This is only applicable to the GroupCreator provider and the Blackboard
consumer. The ajax content being loaded used to be JSP pages. Since all
the JSP pages were moved to /WEB-INF/, they no longer get autogenerated
Servlets, so aren't served up.

As a quick fix, I've created an AjaxServlet that serves up these ajax
pages. It's not very convenient if we have to add new provider or
consumers with ajax functionality though, so it might be best to switch
it up for some other solution in the future.