Fix junk text, IE no bg color.
For some reason, there was an extra "/>" on the options page, removed.

The building block shows up without background colors in IE8 & IE10 for
some reason, turns out to be an IE specific CSS bug where it doesn't
recognize background-color as a property and you need to use just
background, fixed with workaround that specifies a bg color using
background.