Fix error messages not showing up in process.jsp.
Error messages were not showing up if you uploaded a file and it
encountered an error. An exception was shown instead. This was caused by
our UploadMultipartRequestWrapper overriding the original request object.
The wrapper was meant to only wrap a MultipartRequest object despite
having an HttpServletRequest interface. So when the servlet tries to use
it to do HttpServletRequest dispatching, it crashes with an exception.

The exception said something about "JspResourceIncludeUtil is disabled"
which turned out be kind of misleading. The fix was to make sure the
servlet preserves the original HttpServletRequest object for use in
dispatching.

I'm not sure if it's necessary to use the UploadMultipartRequestWrapper
anymore. The problem is that the only Blackboard documentation on
MultipartRequest I have is circa 2001, so now that it's working, I'm
reluctant to touch it in case it disrupts something I don't know about.