Fix exception caused by null group descriptions.
The built-in group tool in BBL was recently updated. This new group tool
doesn't like to have nulls in the group description. A new group object
constructed by the API has null as the default value for group
description. The fix was just to manually set the group description to
an empty string.